requirements.txt
torch~=2.2.2
pillow~=10.3.0
numpy~=1.26.4
scikit-learn~=1.4.1.post1
Flask